Move to real page
I really like the look of the timeline, so it is being moved over to the real page and the old one is being renamed to SomeStories2.html. The constant commits are really just me pushing it out to git so I can see the page on my phone. Yes, there is the browser size change thing but the text is too small to read in that mode. Just pulling it up on the phone works fine